requirements.txt
asgiref==3.10.0
Django==5.2.8
django-filter==25.2
django-restframework==0.0.1
djangorestframework==3.16.1
sqlparse==0.5.3
typing_extensions==4.15.0
